Add files via upload
## Project Overview

This project aims to explore and compare the effectiveness of two different fire detection techniques: Convolutional Neural Network (CNN) MobileNetV2 and Support Vector Machine (SVM). Fire detection is a critical application for ensuring safety in various environments, and this analysis seeks to shed light on the performance and potential of these two methods.

## Motivation

The motivation behind this project is to:

- Assess the accuracy of CNN MobileNetV2 and SVM in identifying fire incidents.
- Understand the trade-offs between deep learning and traditional machine learning approaches.
- Contribute to the field of fire detection and encourage collaborative research.

## Educational Purpose

This project is intended for educational purposes, providing a hands-on example of implementing and comparing different fire detection algorithms. By sharing the code and results, we hope to foster learning and inspire others to explore this vital area of research.

## How to Use

You can replicate and extend this analysis by following the code provided in the Jupyter Notebook files. The dataset is accessible through the Kaggle link mentioned in the 'Data' section.

## Feedback and Contributions

Your feedback and contributions are highly valuable. If you have insights, suggestions, or improvements to share, please don't hesitate to open issues or submit pull requests. Let's collaborate to advance our knowledge of fire detection.